Clean up native intelligence APIs
